my-2-animals-app commit
