+++ .eslintrc.json
@@ -13,5 +13,8 @@
     "sourceType": "module"
   },
   "plugins": ["react"],
-  "rules": {}
+  "rules": {
+    "react/react-in-jsx-scope": "off",
+		"react/prop-types": "off"
+  }
 }
