--- ReactJS-course/my-2-animals-app/.eslintrc.json
@@ -1,20 +1,0 @@
-{
-	"env": {
-		"browser": true,
-		"es2021": true
-	},
-	"extends": [
-		"eslint:recommended",
-		"plugin:react/recommended",
-		"plugin:prettier/recommended"
-	],
-	"parserOptions": {
-		"ecmaVersion": "latest",
-		"sourceType": "module"
-	},
-	"plugins": ["react"],
-	"rules": {
-		"react/react-in-jsx-scope": "off",
-		"react/prop-types": "off"
-	}
-}
