--- ReactJS-course/my-2-animals-app/.prettierignore
@@ -1,4 +1,0 @@
-# Ignore artifacts:
-build
-coverage
-node_modules
