--- ReactJS-course/my-2-animals-app/.prettierrc.json
@@ -1,6 +1,0 @@
-{
-	"printWidth": 85,
-	"singleQuote": false,
-	"tabWidth": 2,
-	"useTabs": true
-}
