--- ReactJS-course/my-2-animals-app/.vscode/settings.json
@@ -1,5 +1,0 @@
-{
-	"editor.codeActionsOnSave": {
-		"source.fixAll.eslint": true
-	}
-}
