--- ReactJS-course/my-2-animals-app/src/AnimalShow.css
@@ -1,22 +1,0 @@
-.animal-wrapper {
-    position: relative;
-    border: 1px solid lightgray;
-    padding: 10px;
-    border-radius: 5px;
-    margin: 10px;
-    box-shadow: 0px 3px 4px lightgray;
-}
-
-.animal {
-    height: 200px;
-    width: 200px;
-    object-fit: contain;
-}
-
-.heart {
-    position: absolute;
-    bottom: 10%;
-    right: 10%;
-    max-height: 200px;
-    max-width: 200px;
-}
