--- ReactJS-course/my-2-animals-app/src/App.css
@@ -1,21 +1,0 @@
-.animals-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-button {
-    background-color: lightgreen;
-    border: 1px solid green;
-    border-radius: 3px;
-    font-size: 20px;
-    padding: 10px;
-    width: 30%;
-}
-
-.animals-list {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: center;
-}
