+++ package.json
@@ -1,12 +1,11 @@
 {
-  "name": "basics-react-app",
+  "name": "my-2-animals-app",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "bulma": "^0.9.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
@@ -46,6 +45,6 @@
     "eslint-plugin-react": "^7.32.2",
     "husky": "^8.0.3",
     "prettier": "^2.8.8",
-    "sass": "^1.62.1"
+    "sass": "^1.63.2"
   }
 }
