+++ src/App.css
@@ -1,0 +1,23 @@
+.animals-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+button {
+  background-color: lightgreen;
+  border: 1px solid green;
+  border-radius: 3px;
+  font-size: 20px;
+  padding: 10px;
+  width: 30%;
+}
+
+.animals-list {
+  width: 75%;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 20px;
+}
